+++ Hypercom/Plugin.cs
@@ -21,7 +21,7 @@ namespace Hypercom
     public sealed class Plugin : BaseUnityPlugin
     {
         public const string Guid = "fulgan.vanguardgalaxy.hypercom";
-        public const string Version = "0.3.2";
+        public const string Version = "0.3.3";
 
         internal static ManualLogSource Log;